depuración metodos innecesarios de validaciñon de reserva clase, ya aparece el botón de cancelar y esa es la validación

+++ lets_talk/app/Http/Responses/estudiante/CancelarClase.php
@@ -31,7 +31,7 @@ class CancelarClase implements Responsable
             try {
                 $claseReservada = Reserva::findOrFail($idClaseReservada->id_reserva);
 
-                if ($claseReservada) {
+                if (isset($claseReservada) && !is_null($claseReservada) && !empty($claseReservada)) {
 
                     $claseCancelada = $claseReservada->forceDelete();
